+++ src/pages/Search.jsx
@@ -1,108 +1,28 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
-import SearchList from '../components/SearchList';
-import { BiSearch } from 'react-icons/bi';
-import { SearchContext } from '../store/search';
-import searchApi from '../services/api';
-import useDebounce from '../hooks/useDebounce';
+import SearchList from './components/SearchList';
 import { useRef } from 'react';
-
-const ArrowDown = 'ArrowDown';
-const ArrowUp = 'ArrowUp';
-const Escape = 'Escape';
+import SearchForm from './components/SearchForm';
 
 const Search = () => {
-  const { searches, setSearches } = useContext(SearchContext);
-  const [searchText, setSearchText] = useState('');
   const [results, setResults] = useState([]);
-  const [index, setIndex] = useState(-1);
-
-  const onSearchSubmit = e => {
-    e.preventDefault();
-    setSearchText('');
-  };
-
-  const onSearchChange = e => {
-    const text = e.target.value;
-    setSearchText(text);
-    setIndex(-1);
-  };
-  const debouncedSearchText = useDebounce(searchText, 200);
-
-  useEffect(() => {
-    if (debouncedSearchText === '') return;
-    if (searches[debouncedSearchText]) {
-      setResults(searches[debouncedSearchText]);
-      return;
-    }
-
-    const getApi = async () => {
-      const results = await searchApi(debouncedSearchText);
-      setSearches(prev => ({ ...prev, [debouncedSearchText]: results.data }));
-      setResults(results.data);
-    };
-    getApi();
-  }, [debouncedSearchText]);
-
-  const autoRef = useRef(null);
   const scrollRef = useRef(null);
-
-  useEffect(() => {
-    console.log('index', index);
-  }, [index]);
-
-  const handleKeyArrow = e => {
-    if (results.length > 0) {
-      switch (e.key) {
-        case ArrowDown:
-          if (index === results.length - 1) {
-            setIndex(0);
-            scrollRef.current?.scrollIntoView({ bebehavior: 'smooth', block: 'center' });
-            return;
-          }
-          setIndex(pre => pre + 1);
-          scrollRef.current?.scrollIntoView({ bebehavior: 'smooth', block: 'center' });
-          if (autoRef.current.childElementCount === index + 1) setIndex(0);
-          break;
-        case ArrowUp:
-          if (index === 0) {
-            setIndex(results.length - 1);
-            scrollRef.current?.scrollIntoView({ bebehavior: 'smooth', block: 'center' });
-            return;
-          }
-
-          setIndex(pre => pre - 1);
-          scrollRef.current?.scrollIntoView({ bebehavior: 'smooth', block: 'center' });
-          break;
-        case Escape: // esc key를 눌렀을때,
-          setResults([]);
-          setIndex(-1);
-          break;
-        default:
-          return;
-      }
-    }
-  };
-
+  const [searchText, setSearchText] = useState('');
+  const [index, setIndex] = useState(-1);
   return (
     <Wrap>
       <TitleWrap>
         <Title>국내 모든 임상시험 검색하고</Title> <Title> 온라인으로 참여하기</Title>
       </TitleWrap>
-
-      <SearchSection onSubmit={onSearchSubmit}>
-        <SearchInput
-          onChange={onSearchChange}
-          value={searchText}
-          placeholder="🔍 검색어를 입력해주세요"
-          ref={autoRef}
-          onKeyDown={handleKeyArrow}
-          autoFocus
-        ></SearchInput>
-        <Button>
-          <BiSearch />
-        </Button>
-      </SearchSection>
+      <SearchForm
+        index={index}
+        setIndex={setIndex}
+        searchText={searchText}
+        setSearchText={setSearchText}
+        scrollRef={scrollRef}
+        results={results}
+        setResults={setResults}
+      />
       {searchText.length !== 0 ? (
         <SearchResult>
           <SearchListTitle>추천 검색어</SearchListTitle>
@@ -147,42 +67,6 @@ const Title = styled.h1`
   font-weight: 700;
   line-height: 1.5em;
   font-size: 2em;
-`;
-const SearchSection = styled.form`
-  display: flex;
-  position: relative;
-  align-items: center;
-  background-color: white;
-  padding: 20px 34px;
-  border-radius: 40px;
-  width: 480px;
-  margin: 0 auto;
-  column-gap: 3em;
-  font-size: 1.7em;
-  box-shadow: 7px 7px 9px #bcd9ed, -7px -7px 9px #d8f9ff;
-`;
-
-const SearchInput = styled.input`
-  display: flex;
-  flex: 1;
-  border: none;
-  outline: none;
-`;
-
-const Button = styled.button`
-  position: absolute;
-  right: 10px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  background-color: #017be8;
-  color: white;
-  font-size: 20px;
-  border: none;
-  cursor: pointer;
 `;
 
 const SearchResult = styled.div`
