+++ src/pages/components/SearchItem.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { highlightIncludedText } from '../utils/func';
+import { highlightIncludedText } from '../../utils/func';
 const SearchItem = React.forwardRef(({ result, searchText, isFocus }, ref) => {
   return isFocus ? (
     <Item ref={ref} isFocus={isFocus}>
@@ -11,10 +11,10 @@ const SearchItem = React.forwardRef(({ result, searchText, isFocus }, ref) => {
   );
 });
 
-export default SearchItem;
-
 const Item = styled.div`
   padding: 6px 0;
   font-size: 14px;
   background-color: ${({ isFocus }) => (isFocus ? '#ddf1ff' : 'white')};
 `;
+
+export default SearchItem;
